requirements.txt
Flask
Flask-SQLAlchemy
pymysql
PyJWT
Werkzeugi
python-dotenv
python-decouple
